requirements.txt
redis==5.0.0
peewee==3.17.5
pyOpenSSL==23.3.0
tornado==6.4.0
uvloop==0.19.0
PyMySQL==1.1.1
cachetools==5.3.3
pycryptodome==3.20.0
psutil==5.9.8
setproctitle==1.3.3
IP2Location==8.10.2
APScheduler==3.10.2
requests==2.32.3
Cython==3.0.10
python_hosts==1.0.6
wheel
setuptools
lamda[full]
frida-tools
mitmproxy>=9.0.0,<=10.2.0
dnspython
httpx[socks]
packaging
paho-mqtt